renamed react files
