--- components/Form/index.js
@@ -1,7 +1,0 @@
-import Button from './Button';
-import Checkbox from './Checkbox';
-import DateInput from './DateInput';
-import SummaryList from './SummaryList';
-import TextInput from './TextInput';
-
-export { Button, Checkbox, DateInput, SummaryList, TextInput };
