+++ components/Form/index.jsx
@@ -1,0 +1,6 @@
+import Button from './Button/Button';
+import Checkbox from './Checkbox/Checkbox';
+import DateInput from './DateInput/DateInput';
+import TextInput from './TextInput/TextInput';
+
+export { Button, Checkbox, DateInput, TextInput };
