+++ components/Steps/result.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Result = props => {
+  return (
+    <>
+      <h2>Result:</h2>
+      <pre>{JSON.stringify(props.formData, null, 2)}</pre>
+    </>
+  );
+};
+
+export default Result;
